+++ package.json
@@ -17,6 +17,7 @@
     "express-validator": "^6.12.0",
     "helmet": "^4.6.0",
     "jsonwebtoken": "^8.5.1",
+    "multer": "^1.4.2",
     "mysql2": "^2.2.5",
     "redis": "^3.1.2"
   },
