+++ routes/index.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-router.get("/", (req, res) => res.send("Dev Server"));
-router.get("/api/hello", (req, res) => res.send("Dev Server"));
 router.use("/api/adminUser", require("./adminUser"));
+router.use("/api/serviceCenter", require("./serviceCenter"));
 
 module.exports = router;
