+++ utils/multer.js
@@ -1,0 +1,22 @@
+const multer = require("multer");
+
+// STORAGE MULTER CONFIG
+let storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+        cb(null, "uploads/");
+    },
+    filename: (req, file, cb) => {
+        cb(null, `${Date.now()}_${file.originalname}`);
+    },
+    fileFilter: (req, file, cb) => {
+        const ext = path.extname(file.originalname)
+        if (ext !== '.jpg' && ext !== '.png' && ext !== '.mp4') {
+            return cb(res.status(400).end('only jpg, png, mp4 is allowed'), false);
+        }
+        cb(null, true)
+    }
+});
+
+const upload = multer({ storage: storage }).single("file");
+
+module.exports = upload;
